+++ app/scripts/main.js
@@ -1,1 +1,3 @@
+import nhsuk_skipLink from './skipLink';
 
+nhsuk_skipLink();